@@ -1,13 +1,23 @@
-import { AppComponent } from './app.component';
-import { CategoryComponent } from './components/category/category.component';
+
 import { Routes, RouterModule } from '@angular/router';
-import { PagesComponent } from './components/shared/pages.component'; 
+import { ForgotPasswordComponent } from './account/forgot-password/forgot-password.component';
+import { LoginComponent } from './account/login/login.component';
 
 const appRoutes: Routes = [
     {
+        path: 'login',
+        component: LoginComponent,
+        data: { title: 'Login' }
+    },
+    {
+        path: 'forgotPassword',
+        component: ForgotPasswordComponent,
+        data: { title: 'Forgot Password' }
+    },
+    {
         path: '',
-        loadChildren: './components/shared/pages.module#PagesModule'
+        loadChildren: () => import('./components/shared/pages.module').then(m => m.PagesModule)
     }
 ];
 
-export const APP_ROUTES = RouterModule.forRoot(appRoutes);
+export const APP_ROUTES = RouterModule.forRoot(appRoutes, { useHash: false });
