@@ -1,5 +1,5 @@
 import { Store } from '../store';
-import { Actions } from 'src/app/enums/actions';
+import { Actions } from 'src/app/utils/enums/actions';
 
 export class StoreSerializer {
   
