@@ -1,6 +1,6 @@
 
 ///All de string messages, url, base configurations goes here
 export class AppSettings {
-    public static API_ENDPOINT = 'https://localhost:44351/api';
+    public static API_ENDPOINT = 'http://localhost:5000/api/v1';
     public static IMAGE_UPLOAD_ERROR = 'Error! Your announcement could not be added!';
 }
