@@ -1,5 +1,5 @@
 import { Resource } from './resource';
-import { Actions } from 'src/app/enums/actions';
+import { Actions } from 'src/app/utils/enums/actions';
 
 export interface Serializer {
     fromJson(json: any): Resource;
