@@ -3,9 +3,7 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
-  urlServer: '/api/',
-  roleAdmin: 'Admin',
+  production: true, 
   AppTitle: 'StoreCoupon',
 };
 
