@@ -1,12 +1,17 @@
 ﻿using Core.Logger.Interface;
 using Core.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Repository.Models.Dtos;
 using System;
 using System.Threading.Tasks;
+using Role = Repository.Repositories.Utils.Role;
 
 namespace GenericProjectBase.Controllers
 {
+    [Route("api/v{version:apiVersion}/storecategories/")]
+    [ApiVersion("2")]
+    [ApiVersion("1")]
     [Route("api/storecategories/")]
     public class CategoryController : Controller
     {
@@ -18,6 +23,10 @@ namespace GenericProjectBase.Controllers
             _loggerManager = loggerManager;
             _categoryService = categoryservice;
         }
+        
+         /// <summary>
+        ///<remarks>Get the categories.</remarks>
+        /// </summary>
         [HttpGet]
         [Route("get")]
         public async Task<IActionResult> Get()
@@ -35,25 +44,39 @@ namespace GenericProjectBase.Controllers
             }
         }
 
+       /// <see cref="id"/> the category model. 
+        /// <summary>
+        /// Get category by guid.
+        /// </summary>
+        /// <returns>
+        /// Category model.
+        /// </returns>
         [HttpGet]
-        [Route("get/{idCategory}")]
-        public async Task<IActionResult> GetById(Guid idCategory)
+        [Route("get/{id}")]
+        public async Task<IActionResult> GetById(Guid id)
         {
             try
             {
-                var query = await _categoryService.GetById(idCategory);
+                var query = await _categoryService.GetById(id);
                 _loggerManager.LogInfo("Owner se obtuvo correctamente");
                 return Ok(query);
             }
             catch (Exception e)
             {
-                _loggerManager.LogError("Ocurrio un error al obtener el category: " + e);
+                _loggerManager.LogError("Ocurrio un error al obtener el category Id: " +id + e);
                 return StatusCode(500);
             }
         }
 
-        [HttpPost]
+        /// <summary>
+        /// Save the categories
+        /// <see cref="CategoryStoreDto"/> the sub category model. 
+        /// </summary>
+        /// <param name="category"></param>
+        //[Authorize(Roles = Role.Admin)]
         [Route("save")]
+        [MapToApiVersion("2")] 
+        [HttpPost]
         public async Task<IActionResult> Add([FromBody] CategoryStoreDto category)
         {
             try
@@ -71,13 +94,19 @@ namespace GenericProjectBase.Controllers
 
         }
 
+        /// <summary>
+        /// Delete Category by Id.
+        /// </summary>
+        /// <returns>
+        /// Category deleted.
+        /// </returns>
         [HttpDelete]
-        [Route("delete/{idCategory}")]
-        public async Task<IActionResult> DeleteById(Guid idCategory)
+        [Route("delete/{id}")]
+        public async Task<IActionResult> DeleteById(Guid id)
         {
             try
             {
-                var query = await _categoryService.DeleteById(idCategory);
+                var query = await _categoryService.DeleteById(id);
                 _loggerManager.LogInfo("El category fue eliminado correctamente");
                 return Ok(query);
             }
@@ -88,6 +117,11 @@ namespace GenericProjectBase.Controllers
             }
         }
 
+        /// <summary>
+        /// Update the category.
+        /// <see cref="CategoryStoreDto"/> the category model. 
+        /// </summary>
+        /// <param name="category"></param>
         [HttpPut]
         [Route("update")]
         public async Task<IActionResult> Update([FromBody] CategoryStoreDto category)
